Function esp_idf_bindgen::esp_event_handler_unregister

pub unsafe extern "C" fn esp_event_handler_unregister(
    event_base: esp_event_base_t, 
    event_id: i32, 
    event_handler: esp_event_handler_t
) -> esp_err_t

@brief Unregister a handler with the system event loop (legacy).

@note This function is obsolete and will be deprecated soon, please use esp_event_handler_instance_unregister() instead.

This function can be used to unregister a handler so that it no longer gets called during dispatch. Handlers can be unregistered for either: (1) specific events, (2) all events of a certain event base, or (3) all events known by the system event loop

specific events: specify exact event_base and event_id

all events of a certain base: specify exact event_base and use ESP_EVENT_ANY_ID as the event_id

all events known by the loop: use ESP_EVENT_ANY_BASE for event_base and ESP_EVENT_ANY_ID as the event_id

This function ignores unregistration of handlers that has not been previously registered.

@param[in] event_base the base of the event with which to unregister the handler @param[in] event_id the id of the event with which to unregister the handler @param[in] event_handler the handler to unregister

@return ESP_OK success @return ESP_ERR_INVALID_ARG invalid combination of event base and event id @return others fail