Function esp_idf_bindgen::esp_event_loop_set_cb

pub unsafe extern "C" fn esp_event_loop_set_cb(
    cb: system_event_cb_t, 
    ctx: *mut c_void
) -> system_event_cb_t

@brief Set application specified event callback function

@note This API is part of the legacy event system. New code should use event library API in esp_event.h

@attention 1. If cb is NULL, means application don't need to handle If cb is not NULL, it will be call when an event is received, after the default event callback is completed

@param cb application callback function @param ctx argument to be passed to callback

@return old callback