Function esp_idf_bindgen::esp_netif_attach

pub unsafe extern "C" fn esp_netif_attach(
    esp_netif: *mut esp_netif_t, 
    driver_handle: esp_netif_iodriver_handle
) -> esp_err_t

@brief Attaches esp_netif instance to the io driver handle

Calling this function enables connecting specific esp_netif object with already initialized io driver to update esp_netif object with driver specific configuration (i.e. calls post_attach callback, which typically sets io driver callbacks to esp_netif instance and starts the driver)

@param[inout] esp_netif pointer to esp_netif object to be attached @param[in] driver_handle pointer to the driver handle @return - ESP_OK on success - ESP_ERR_ESP_NETIF_DRIVER_ATTACH_FAILED if driver's pot_attach callback failed