Function esp_idf_bindgen::esp_ota_end

pub unsafe extern "C" fn esp_ota_end(handle: esp_ota_handle_t) -> esp_err_t

@brief Finish OTA update and validate newly written app image.

@param handle Handle obtained from esp_ota_begin().

@note After calling esp_ota_end(), the handle is no longer valid and any memory associated with it is freed (regardless of result).

@return

ESP_OK: Newly written OTA app image is valid.

ESP_ERR_NOT_FOUND: OTA handle was not found.

ESP_ERR_INVALID_ARG: Handle was never written to.

ESP_ERR_OTA_VALIDATE_FAILED: OTA image is invalid (either not a valid app image, or - if secure boot is enabled - signature failed to verify.)

ESP_ERR_INVALID_STATE: If flash encryption is enabled, this result indicates an internal error writing the final encrypted bytes to flash.