Function esp_idf_bindgen::esp_ota_mark_app_invalid_rollback_and_reboot

pub unsafe extern "C" fn esp_ota_mark_app_invalid_rollback_and_reboot(
) -> esp_err_t

@brief This function is called to roll back to the previously workable app with reboot.

If rollback is successful then device will reset else API will return with error code. Checks applications on a flash drive that can be booted in case of rollback. If the flash does not have at least one app (except the running app) then rollback is not possible. @return

ESP_FAIL: if not successful.

ESP_ERR_OTA_ROLLBACK_FAILED: The rollback is not possible due to flash does not have any apps.