Function esp_idf_bindgen::esp_partition_get_sha256

pub unsafe extern "C" fn esp_partition_get_sha256(
    partition: *const esp_partition_t, 
    sha_256: *mut u8
) -> esp_err_t

@brief Get SHA-256 digest for required partition.

For apps with SHA-256 appended to the app image, the result is the appended SHA-256 value for the app image content. The hash is verified before returning, if app content is invalid then the function returns ESP_ERR_IMAGE_INVALID. For apps without SHA-256 appended to the image, the result is the SHA-256 of all bytes in the app image. For other partition types, the result is the SHA-256 of the entire partition.

@param[in] partition Pointer to info for partition containing app or data. (fields: address, size and type, are required to be filled). @param[out] sha_256 Returned SHA-256 digest for a given partition.

@return - ESP_OK: In case of successful operation. - ESP_ERR_INVALID_ARG: The size was 0 or the sha_256 was NULL. - ESP_ERR_NO_MEM: Cannot allocate memory for sha256 operation. - ESP_ERR_IMAGE_INVALID: App partition doesn't contain a valid app image. - ESP_FAIL: An allocation error occurred.