Function esp_idf_bindgen::esp_partition_verify

pub unsafe extern "C" fn esp_partition_verify(
    partition: *const esp_partition_t
) -> *const esp_partition_t

@brief Verify partition data

Given a pointer to partition data, verify this partition exists in the partition table (all fields match.)

This function is also useful to take partition data which may be in a RAM buffer and convert it to a pointer to the permanent partition data stored in flash.

Pointers returned from this function can be compared directly to the address of any pointer returned from esp_partition_get(), as a test for equality.

@param partition Pointer to partition data to verify. Must be non-NULL. All fields of this structure must match the partition table entry in flash for this function to return a successful match.

@return

If partition not found, returns NULL.

If found, returns a pointer to the esp_partition_t structure in flash. This pointer is always valid for the lifetime of the application.