Function esp_idf_bindgen::esp_set_deep_sleep_wake_stub

pub unsafe extern "C" fn esp_set_deep_sleep_wake_stub(
    new_stub: esp_deep_sleep_wake_stub_fn_t
)

@brief Install a new stub at runtime to run on wake from deep sleep

If implementing esp_wake_deep_sleep() then it is not necessary to call this function.

However, it is possible to call this function to substitute a different deep sleep stub. Any function used as a deep sleep stub must be marked RTC_IRAM_ATTR, and must obey the same rules given for esp_wake_deep_sleep().