Function esp_idf_bindgen::esp_wifi_ap_get_sta_list

pub unsafe extern "C" fn esp_wifi_ap_get_sta_list(
    sta: *mut wifi_sta_list_t
) -> esp_err_t

@brief Get STAs associated with soft-AP

@attention SSC only API

@param[out] sta station list ap can get the connected sta's phy mode info through the struct member phy_11b，phy_11g，phy_11n，phy_lr in the wifi_sta_info_t struct. For example, phy_11b = 1 imply that sta support 802.11b mode

@return

ESP_OK: succeed

ESP_ERR_WIFI_NOT_INIT: WiFi is not initialized by esp_wifi_init

ESP_ERR_INVALID_ARG: invalid argument

ESP_ERR_WIFI_MODE: WiFi mode is wrong

ESP_ERR_WIFI_CONN: WiFi internal error, the station/soft-AP control block is invalid