Function esp_idf_bindgen::heap_caps_aligned_alloc

pub unsafe extern "C" fn heap_caps_aligned_alloc(
    alignment: size_t, 
    size: size_t, 
    caps: c_int
) -> *mut c_void

@brief Allocate a aligned chunk of memory which has the given capabilities

Equivalent semantics to libc aligned_alloc(), for capability-aware memory. @param alignment How the pointer received needs to be aligned must be a power of two @param size Size, in bytes, of the amount of memory to allocate @param caps Bitwise OR of MALLOC_CAP_* flags indicating the type of memory to be returned

@return A pointer to the memory allocated on success, NULL on failure

@note Any memory allocated with heaps_caps_aligned_alloc() MUST be freed with heap_caps_aligned_free() and CANNOT be passed to free()