Function esp_idf_bindgen::heap_caps_check_integrity

pub unsafe extern "C" fn heap_caps_check_integrity(
    caps: u32, 
    print_errors: bool
) -> bool

@brief Check integrity of all heaps with the given capabilities.

Calls multi_heap_check on all heaps which share the given capabilities. Optionally print errors if the heaps are corrupt.

See also heap_caps_check_integrity_all to check all heap memory in the system and heap_caps_check_integrity_addr to check memory around a single address.

@param caps Bitwise OR of MALLOC_CAP_* flags indicating the type of memory @param print_errors Print specific errors if heap corruption is found.

@return True if all heaps are valid, False if at least one heap is corrupt.