Function esp_idf_bindgen::heap_caps_check_integrity_all

pub unsafe extern "C" fn heap_caps_check_integrity_all(
    print_errors: bool
) -> bool

@brief Check integrity of all heap memory in the system.

Calls multi_heap_check on all heaps. Optionally print errors if heaps are corrupt.

Calling this function is equivalent to calling heap_caps_check_integrity with the caps argument set to MALLOC_CAP_INVALID.

@param print_errors Print specific errors if heap corruption is found.

@return True if all heaps are valid, False if at least one heap is corrupt.