Function esp_idf_bindgen::i2c_slave_read_buffer

pub unsafe extern "C" fn i2c_slave_read_buffer(
    i2c_num: i2c_port_t, 
    data: *mut u8, 
    max_size: size_t, 
    ticks_to_wait: TickType_t
) -> c_int

@brief I2C slave read data from internal buffer. When I2C slave receive data, isr will copy received data from hardware rx fifo to internal ringbuffer. Then users can read from internal ringbuffer. @note Only call this function in I2C slave mode

@param i2c_num I2C port number @param data data pointer to accept data from internal buffer @param max_size Maximum data size to read @param ticks_to_wait Maximum waiting ticks

@return - ESP_FAIL(-1) Parameter error - Others(>=0) The number of data bytes that read from I2C slave buffer.