Function esp_idf_bindgen::nvs_flash_secure_init_partition

pub unsafe extern "C" fn nvs_flash_secure_init_partition(
    partition_label: *const c_char, 
    cfg: *mut nvs_sec_cfg_t
) -> esp_err_t

@brief Initialize NVS flash storage for the specified partition.

@param[in] partition_label Label of the partition. Note that internally a reference to passed value is kept and it should be accessible for future operations

@param[in] cfg Security configuration (keys) to be used for NVS encryption/decryption. If cfg is null, no encryption/decryption is used. @return - ESP_OK if storage was successfully initialized. - ESP_ERR_NVS_NO_FREE_PAGES if the NVS storage contains no empty pages (which may happen if NVS partition was truncated) - ESP_ERR_NOT_FOUND if specified partition is not found in the partition table - one of the error codes from the underlying flash storage driver