Function esp_idf_bindgen::rtc_clk_init

pub unsafe extern "C" fn rtc_clk_init(cfg: rtc_clk_config_t)

Initialize clocks and set CPU frequency

If cfg.xtal_freq is set to RTC_XTAL_FREQ_AUTO, this function will attempt to auto detect XTAL frequency. Auto detection is performed by comparing XTAL frequency with the frequency of internal 8MHz oscillator. Note that at high temperatures the frequency of the internal 8MHz oscillator may drift enough for auto detection to be unreliable. Auto detection code will attempt to distinguish between 26MHz and 40MHz crystals. 24 MHz crystals are not supported by auto detection code. If XTAL frequency can not be auto detected, this 26MHz frequency will be used.

@param cfg clock configuration as rtc_clk_config_t