Function esp_idf_bindgen::touch_pad_filter_start

pub unsafe extern "C" fn touch_pad_filter_start(
    filter_period_ms: u32
) -> esp_err_t

@brief start touch pad filter function This API will start a filter to process the noise in order to prevent false triggering when detecting slight change of capacitance. Need to call touch_pad_filter_start before all touch filter APIs

@note This filter uses FreeRTOS timer, which is dispatched from a task with priority 1 by default on CPU 0. So if some application task with higher priority takes a lot of CPU0 time, then the quality of data obtained from this filter will be affected. You can adjust FreeRTOS timer task priority in menuconfig. @param filter_period_ms filter calibration period, in ms @return - ESP_OK Success - ESP_ERR_INVALID_ARG parameter error - ESP_ERR_NO_MEM No memory for driver - ESP_ERR_INVALID_STATE driver state error