Function esp_idf_bindgen::uart_set_rx_full_threshold

pub unsafe extern "C" fn uart_set_rx_full_threshold(
    uart_num: uart_port_t, 
    threshold: c_int
) -> esp_err_t

@brief Set uart threshold value for RX fifo full @note If application is using higher baudrate and it is observed that bytes in hardware RX fifo are overwritten then this threshold can be reduced

@param uart_num UART_NUM_0, UART_NUM_1 or UART_NUM_2 @param threshold Threshold value above which RX fifo full interrupt is generated

@return - ESP_OK Success - ESP_ERR_INVALID_ARG Parameter error - ESP_ERR_INVALID_STATE Driver is not installed