Function esp_idf_bindgen::uxTaskGetStackHighWaterMark

pub unsafe extern "C" fn uxTaskGetStackHighWaterMark(
    xTask: TaskHandle_t
) -> UBaseType_t

Returns the high water mark of the stack associated with xTask.

INCLUDE_uxTaskGetStackHighWaterMark must be set to 1 in FreeRTOSConfig.h for this function to be available.

High water mark is the minimum free stack space there has been (in bytes rather than words as found in vanilla FreeRTOS) since the task started. The smaller the returned number the closer the task has come to overflowing its stack.

@param xTask Handle of the task associated with the stack to be checked. Set xTask to NULL to check the stack of the calling task.

@return The smallest amount of free stack space there has been (in bytes rather than words as found in vanilla FreeRTOS) since the task referenced by xTask was created.