Function esp_idf_bindgen::vTaskStartScheduler

pub unsafe extern "C" fn vTaskStartScheduler()

@cond */ Starts the real time kernel tick processing.

After calling the kernel has control over which tasks are executed and when.

See the demo application file main.c for an example of creating tasks and starting the kernel.

Example usage: @code{c} void vAFunction( void ) { // Create at least one task before starting the kernel. xTaskCreate( vTaskCode, "NAME", STACK_SIZE, NULL, tskIDLE_PRIORITY, NULL );

// Start the real time kernel with preemption. vTaskStartScheduler ();

// Will not get here unless a task calls vTaskEndScheduler () } @endcode

\ingroup SchedulerControl