Function esp_idf_bindgen::xQueueAddToSet

pub unsafe extern "C" fn xQueueAddToSet(
    xQueueOrSemaphore: QueueSetMemberHandle_t, 
    xQueueSet: QueueSetHandle_t
) -> BaseType_t

Adds a queue or semaphore to a queue set that was previously created by a call to xQueueCreateSet().

See FreeRTOS/Source/Demo/Common/Minimal/QueueSet.c for an example using this function.

Note 1: A receive (in the case of a queue) or take (in the case of a semaphore) operation must not be performed on a member of a queue set unless a call to xQueueSelectFromSet() has first returned a handle to that set member.

@param xQueueOrSemaphore The handle of the queue or semaphore being added to the queue set (cast to an QueueSetMemberHandle_t type).

@param xQueueSet The handle of the queue set to which the queue or semaphore is being added.

@return If the queue or semaphore was successfully added to the queue set then pdPASS is returned. If the queue could not be successfully added to the queue set because it is already a member of a different queue set then pdFAIL is returned.