Function esp_idf_bindgen::xQueueGenericReceive

pub unsafe extern "C" fn xQueueGenericReceive(
    xQueue: QueueHandle_t, 
    pvBuffer: *mut c_void, 
    xTicksToWait: TickType_t, 
    xJustPeek: BaseType_t
) -> BaseType_t

It is preferred that the macro xQueueReceive() be used rather than calling this function directly.

Receive an item from a queue. The item is received by copy so a buffer of adequate size must be provided. The number of bytes copied into the buffer was defined when the queue was created.

This function must not be used in an interrupt service routine. See xQueueReceiveFromISR for an alternative that can.

@param xQueue The handle to the queue from which the item is to be received.

@param pvBuffer Pointer to the buffer into which the received item will be copied.

@param xTicksToWait The maximum amount of time the task should block waiting for an item to receive should the queue be empty at the time of the call. The time is defined in tick periods so the constant portTICK_PERIOD_MS should be used to convert to real time if this is required. xQueueGenericReceive() will return immediately if the queue is empty and xTicksToWait is 0.

@param xJustPeek When set to true, the item received from the queue is not actually removed from the queue - meaning a subsequent call to xQueueReceive() will return the same item. When set to false, the item being received from the queue is also removed from the queue.

@return pdTRUE if an item was successfully received from the queue, otherwise pdFALSE.

Example usage: @code{c} struct AMessage { char ucMessageID; char ucData[ 20 ]; } xMessage;

QueueHandle_t xQueue;

// Task to create a queue and post a value. void vATask( void *pvParameters ) { struct AMessage *pxMessage;

// Create a queue capable of containing 10 pointers to AMessage structures.
// These should be passed by pointer as they contain a lot of data.
xQueue = xQueueCreate( 10, sizeof( struct AMessage * ) );
if( xQueue == 0 )
{
	// Failed to create the queue.
}

// ...

// Send a pointer to a struct AMessage object.  Don't block if the
// queue is already full.
pxMessage = & xMessage;
xQueueSend( xQueue, ( void * ) &pxMessage, ( TickType_t ) 0 );

// ... Rest of task code.

}

// Task to receive from the queue. void vADifferentTask( void *pvParameters ) { struct AMessage *pxRxedMessage;

if( xQueue != 0 )
{
	// Receive a message on the created queue.  Block for 10 ticks if a
	// message is not immediately available.
	if( xQueueGenericReceive( xQueue, &( pxRxedMessage ), ( TickType_t ) 10 ) )
	{
		// pcRxedMessage now points to the struct AMessage variable posted
		// by vATask.
	}
}

// ... Rest of task code.

} @endcode \ingroup QueueManagement