Function esp_idf_bindgen::xQueueRemoveFromSet

pub unsafe extern "C" fn xQueueRemoveFromSet(
    xQueueOrSemaphore: QueueSetMemberHandle_t, 
    xQueueSet: QueueSetHandle_t
) -> BaseType_t

Removes a queue or semaphore from a queue set. A queue or semaphore can only be removed from a set if the queue or semaphore is empty.

See FreeRTOS/Source/Demo/Common/Minimal/QueueSet.c for an example using this function.

@param xQueueOrSemaphore The handle of the queue or semaphore being removed from the queue set (cast to an QueueSetMemberHandle_t type).

@param xQueueSet The handle of the queue set in which the queue or semaphore is included.

@return If the queue or semaphore was successfully removed from the queue set then pdPASS is returned. If the queue was not in the queue set, or the queue (or semaphore) was not empty, then pdFAIL is returned.