Function esp_idf_bindgen::xRingbufferGetMaxItemSize

pub unsafe extern "C" fn xRingbufferGetMaxItemSize(
    xRingbuffer: RingbufHandle_t
) -> size_t

@brief Get maximum size of an item that can be placed in the ring buffer

This function returns the maximum size an item can have if it was placed in an empty ring buffer.

@param[in] xRingbuffer Ring buffer to query

@note The max item size for a no-split buffer is limited to ((buffer_size/2)-header_size). This limit is imposed so that an item of max item size can always be sent to the an empty no-split buffer regardless of the internal positions of the buffer's read/write/free pointers.

@return Maximum size, in bytes, of an item that can be placed in a ring buffer.