Struct esp_idf_bindgen::wpa_crypto_funcs_t

#[repr(C)]pub struct wpa_crypto_funcs_t {
    pub size: u32,
    pub version: u32,
    pub aes_wrap: esp_aes_wrap_t,
    pub aes_unwrap: esp_aes_unwrap_t,
    pub hmac_sha256_vector: esp_hmac_sha256_vector_t,
    pub sha256_prf: esp_sha256_prf_t,
    pub hmac_md5: esp_hmac_md5_t,
    pub hamc_md5_vector: esp_hmac_md5_vector_t,
    pub hmac_sha1: esp_hmac_sha1_t,
    pub hmac_sha1_vector: esp_hmac_sha1_vector_t,
    pub sha1_prf: esp_sha1_prf_t,
    pub sha1_vector: esp_sha1_vector_t,
    pub pbkdf2_sha1: esp_pbkdf2_sha1_t,
    pub rc4_skip: esp_rc4_skip_t,
    pub md5_vector: esp_md5_vector_t,
    pub aes_encrypt: esp_aes_encrypt_t,
    pub aes_encrypt_init: esp_aes_encrypt_init_t,
    pub aes_encrypt_deinit: esp_aes_encrypt_deinit_t,
    pub aes_decrypt: esp_aes_decrypt_t,
    pub aes_decrypt_init: esp_aes_decrypt_init_t,
    pub aes_decrypt_deinit: esp_aes_decrypt_deinit_t,
    pub omac1_aes_128: esp_omac1_aes_128_t,
    pub ccmp_decrypt: esp_ccmp_decrypt_t,
    pub ccmp_encrypt: esp_ccmp_encrypt_t,
}

@brief The crypto callback function structure used when do station security connect. The structure can be set as software crypto or the crypto optimized by ESP32 hardware.

Fields
size: u32version: u32aes_wrap: esp_aes_wrap_t

< station connect function used when send EAPOL frame

aes_unwrap: esp_aes_unwrap_t

< station connect function used when decrypt key data

hmac_sha256_vector: esp_hmac_sha256_vector_t

< station connect function used when check MIC

sha256_prf: esp_sha256_prf_t

< station connect function used when check MIC

hmac_md5: esp_hmac_md5_thamc_md5_vector: esp_hmac_md5_vector_thmac_sha1: esp_hmac_sha1_thmac_sha1_vector: esp_hmac_sha1_vector_tsha1_prf: esp_sha1_prf_tsha1_vector: esp_sha1_vector_tpbkdf2_sha1: esp_pbkdf2_sha1_trc4_skip: esp_rc4_skip_tmd5_vector: esp_md5_vector_taes_encrypt: esp_aes_encrypt_taes_encrypt_init: esp_aes_encrypt_init_taes_encrypt_deinit: esp_aes_encrypt_deinit_taes_decrypt: esp_aes_decrypt_taes_decrypt_init: esp_aes_decrypt_init_taes_decrypt_deinit: esp_aes_decrypt_deinit_tomac1_aes_128: esp_omac1_aes_128_tccmp_decrypt: esp_ccmp_decrypt_tccmp_encrypt: esp_ccmp_encrypt_t